+++ CHANGELOG.md
@@ -9,6 +9,7 @@ description: |-
 
 #### v0.9.1 (unreleased)
 - Fix panic on VKCS Kubernetes API error in vkcs_kubernetes_security_policy
+- Update Let's Encrypt certificate issuance process in vkcs_cdn_resource
 
 #### v0.9.0
 - Add vkcs_cloud_monitoring resource